refactor(indexeddb): expose safe encode trait even when e2e-encryption feature is not enabled
