+++ crates/matrix-sdk-indexeddb/src/lib.rs
@@ -12,7 +12,6 @@ mod error;
 mod event_cache_store;
 #[cfg(feature = "media-store")]
 mod media_store;
-#[cfg(feature = "e2e-encryption")]
 mod serializer;
 #[cfg(feature = "state-store")]
 mod state_store;
